chore: Moves Switch to the components folder (#14135)
